feb11/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.4
Django==6.0.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
django-unfold==0.78.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
pycparser==3.0
PyJWT==2.11.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
